Create calculatePercentile method
